requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
jupyter==1.0.0
imbalanced-learn==0.11.0
joblib==1.3.2
faker==19.6.2
tqdm==4.66.1

# Real-time streaming dependencies
kafka-python==2.0.2
redis==4.5.4
psycopg2-binary==2.9.6
flask==2.3.2
requests==2.28.2